changes before me

+++ ldm/models/diffusion/ddim.py
@@ -12,7 +12,7 @@ from ldm.modules.diffusionmodules.util import make_ddim_sampling_parameters, mak
 class DDIMSampler(object):
     def __init__(self, model, schedule="linear", **kwargs):
         super().__init__()
-        self.model = model
+        self.model = model.half()
         self.ddpm_num_timesteps = model.num_timesteps
         self.schedule = schedule
